requirements.txt
torch>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
plotly>=5.10.0
pyyaml>=6.0
requests>=2.31.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
prometheus-client>=0.17.0
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.12.0
gunicorn>=21.0.0
